Fix loading of binary_compressed organized PCD
The patch fixes an issue where only the first line of data of an
organized PCD was loaded.

The size of an organised PCD is defined by both its width and its
height, and the amount of points to load is defined by the number of
points and not solely by its width.